Changed the vcf tag prefix t into di (except that tHap is changed into delinsHap for better clarity at first glance). Generate more user-friendly error message when the sequence records in the FASTA and VCF files do not match with each other. Put delinsVariantCaller info more at the end in the VCF header to better maintain compatiblity with very old versions of pysam. Now generate (*RDM and *RDm) files. Updated to version 0.1.3. Fixed a compiling error due to arrays not downgraded to primitiives yet. Fixed a compiler warning when the FASTA and VCF files have different template sequences.
Please note that this commit is not backward compatible due to changes to the VCF tags. The reason for such changes is due to the fact that this program is not just used for tumor sequencing data. For example, this program may be run on healthy germline variants and still generate biologically meaningful delins variants. Hence, the prefix t is very misleading. Additionally, the ADR can be changed into AD for better resemblance with the naming in VCF specs and GATK.
Please also note that many VCF tags have Number=A not Number=R mainly due to the pecularity of delins variant ALT alleles compared with the wild-type REF alleles.

Updated to version 0.1.3.